Make GUI OOP
