+++ src/main/java/compal/logic/commands/DoneCommand.java
@@ -2,7 +2,7 @@ package compal.logic.commands;
 
 import compal.commons.Compal;
 import compal.logic.parser.CommandParser;
-import compal.tasks.TaskList;
+import compal.model.tasks.TaskList;
 
 import java.util.Scanner;
