+++ src/main/java/compal/logic/commands/EventCommand.java
@@ -2,9 +2,9 @@ package compal.logic.commands;
 
 import compal.commons.Compal;
 import compal.logic.parser.CommandParser;
-import compal.tasks.Event;
-import compal.tasks.Task;
-import compal.tasks.TaskList;
+import compal.model.tasks.Event;
+import compal.model.tasks.Task;
+import compal.model.tasks.TaskList;
 
 import java.util.Scanner;
