+++ src/main/java/compal/logic/commands/FixedDurationCommand.java
@@ -2,9 +2,9 @@ package compal.logic.commands;
 
 import compal.commons.Compal;
 import compal.logic.parser.CommandParser;
-import compal.tasks.FixedDurationTask;
-import compal.tasks.Task;
-import compal.tasks.TaskList;
+import compal.model.tasks.FixedDurationTask;
+import compal.model.tasks.Task;
+import compal.model.tasks.TaskList;
 
 import java.util.Scanner;
