+++ src/main/java/compal/logic/commands/RecurTaskCommand.java
@@ -2,9 +2,9 @@ package compal.logic.commands;
 
 import compal.commons.Compal;
 import compal.logic.parser.CommandParser;
-import compal.tasks.RecurringTask;
-import compal.tasks.Task;
-import compal.tasks.TaskList;
+import compal.model.tasks.RecurringTask;
+import compal.model.tasks.Task;
+import compal.model.tasks.TaskList;
 
 import java.text.ParseException;
 import java.text.SimpleDateFormat;
