+++ src/main/java/compal/model/tasks/FixedDurationTask.java
@@ -1,4 +1,4 @@
-package compal.tasks;
+package compal.model.tasks;
 
 /**
  * Represents event task type with a starting date and time, as well as a fixed duration.
